+++ samples/CannedDatabase/Resources/app.js
@@ -11,14 +11,14 @@ var TiTouchDB = require('com.obscure.titouchdb');
  * technique would not be suitable for large datasets.
  */
 function generateDB() {
-  var db = TiTouchDB.databaseManager.createDatabaseNamed('elements');
+  var db = TiTouchDB.databaseManager.getDatabase('elements');
   var elements = [[1, "Hydrogen", "H"], [2, "Helium", "He"], [3, "Lithium", "Li"], [4, "Beryllium", "Be"], [5, "Boron", "B"], [6, "Carbon", "C"], [7, "Nitrogen", "N"], [8, "Oxygen", "O"], [9, "Fluorine", "F"], [10, "Neon", "Ne"], [11, "Sodium", "Na"], [12, "Magnesium", "Mg"], [13, "Aluminium", "Al", "http://0.tqn.com/d/chemistry/1/6/q/V/1/Aluminium.jpg"], [14, "Silicon", "Si"], [15, "Phosphorus", "P"], [16, "Sulfur", "S"], [17, "Chlorine", "Cl"], [18, "Argon", "Ar"], [19, "Potassium", "K"], [20, "Calcium", "Ca"], [21, "Scandium", "Sc"], [22, "Titanium", "Ti"], [23, "Vanadium", "V"], [24, "Chromium", "Cr"], [25, "Manganese", "Mn"], [26, "Iron", "Fe"], [27, "Cobalt", "Co", "http://0.tqn.com/d/chemistry/1/6/I/Z/1/cobalt.jpg"], [28, "Nickel", "Ni"], [29, "Copper", "Cu"], [30, "Zinc", "Zn"], [31, "Gallium", "Ga", "http://0.tqn.com/d/chemistry/1/6/H/Q/gallium.jpg"], [32, "Germanium", "Ge"], [33, "Arsenic", "As"], [34, "Selenium", "Se"], [35, "Bromine", "Br"], [36, "Krypton", "Kr"], [37, "Rubidium", "Rb"], [38, "Strontium", "Sr"], [39, "Yttrium", "Y"], [40, "Zirconium", "Zr"], [41, "Niobium", "Nb"], [42, "Molybdenum", "Mo"], [43, "Technetium", "Tc"], [44, "Ruthenium", "Ru"], [45, "Rhodium", "Rh"], [46, "Palladium", "Pd"], [47, "Silver", "Ag"], [48, "Cadmium", "Cd"], [49, "Indium", "In"], [50, "Tin", "Sn"], [51, "Antimony", "Sb"], [52, "Tellurium", "Te"], [53, "Iodine", "I"], [54, "Xenon", "Xe"], [55, "Caesium", "Cs"], [56, "Barium", "Ba"], [71, "Lutetium", "Lu"], [72, "Hafnium", "Hf"], [73, "Tantalum", "Ta"], [74, "Tungsten", "W"], [75, "Rhenium", "Re"], [76, "Osmium", "Os"], [77, "Iridium", "Ir"], [78, "Platinum", "Pt"], [79, "Gold", "Au"], [80, "Mercury", "Hg"], [81, "Thallium", "Tl"], [82, "Lead", "Pb"], [83, "Bismuth", "Bi", "http://0.tqn.com/d/chemistry/1/6/m/Q/bismuth.jpg"], [84, "Polonium", "Po"], [85, "Astatine", "At"], [86, "Radon", "Rn"], [87, "Francium", "Fr"], [88, "Radium", "Ra"], [103, "Lawrencium", "Lr"], [104, "Rutherfordium", "Rf"], [105, "Dubnium", "Db"], [106, "Seaborgium", "Sg"], [107, "Bohrium", "Bh"], [108, "Hassium", "Hs"], [109, "Meitnerium", "Mt"], [110, "Darmstadtium", "Ds"], [111, "Roentgenium", "Rg"], [112, "Copernicium", "Cn"], [113, "Ununtrium", "Uut"], [114, "Ununquadium", "Uuq"], [115, "Ununpentium", "Uup"], [116, "Ununhexium", "Uuh"], [117, "Ununseptium", "Uus"], [118, "Ununoctium", "Uuo"], [57, "Lanthanum", "La"], [58, "Cerium", "Ce"], [59, "Praseodymium", "Pr"], [60, "Neodymium", "Nd"], [61, "Promethium", "Pm"], [62, "Samarium", "Sm"], [63, "Europium", "Eu"], [64, "Gadolinium", "Gd"], [65, "Terbium", "Tb"], [66, "Dysprosium", "Dy"], [67, "Holmium", "Ho"], [68, "Erbium", "Er"], [69, "Thulium", "Tm"], [70, "Ytterbium", "Yb"], [89, "Actinium", "Ac"], [90, "Thorium", "Th"], [91, "Protactinium", "Pa"], [92, "Uranium", "U"], [93, "Neptunium", "Np"], [94, "Plutonium", "Pu"], [95, "Americium", "Am"], [96, "Curium", "Cm"], [97, "Berkelium", "Bk"], [98, "Californium", "Cf"], [99, "Einsteinium", "Es"], [100, "Fermium", "Fm"], [101, "Mendelevium", "Md"], [102, "Nobelium", "No"]];
   
   function saveAttachment(doc, url) {
-    ++attcount''
+    ++attcount;
     var client = Ti.Network.createHTTPClient({
       onload: function(e) {
-        var rev = doc.newRevision();
+        var rev = doc.createRevision();
         rev.addAttachment('image.jpg', 'image/jpeg', client.responseData);
         rev.save();
       }
@@ -30,7 +30,7 @@ function generateDB() {
   var attcount = 0;
   for (i=0; i < elements.length; i++) {
     var e = elements[i];
-    var doc = db.untitledDocument();
+    var doc = db.createDocument();
     doc.putProperties({
       type: 'element',
       atomic_number: e[0],
@@ -107,6 +107,15 @@ function copyDatabaseFiles() {
   copydir(srcdir, destdir);
 }
 
+function installDatabase() {
+  var basedir = Ti.Filesystem.getFile(Ti.Filesystem.resourcesDirectory, 'assets', 'CouchbaseLite').path;
+  var dbfile = [basedir, 'elements.cblite'].join(Ti.Filesystem.separator);
+  var attdir = [basedir, 'elements attachments'].join(Ti.Filesystem.separator);
+  if (!TiTouchDB.databaseManager.replaceDatabase('elements', dbfile, attdir)) {
+    throw 'could not install elements database';
+  }  
+}
+
 var win = Ti.UI.createWindow({
   backgroundColor: 'white'
 });
@@ -116,22 +125,22 @@ var tableView = Ti.UI.createTableView({
 win.add(tableView);
 
 win.addEventListener('open', function(e) {
-  var db = TiTouchDB.databaseManager.databaseNamed('elements');
+  var db = TiTouchDB.databaseManager.getExistingDatabase('elements');
   if (!db) {
-    copyDatabaseFiles();
-    db = TiTouchDB.databaseManager.databaseNamed('elements');
+    installDatabase();
+    db = TiTouchDB.databaseManager.getExistingDatabase('elements');
     if (!db) {
-      alert("Error copying database files!");
+      alert("Error creating database!");
       return;
     }
     Ti.API.info("copied database files");
   }
 
-  var query = db.queryAllDocuments();
+  var query = db.createAllDocumentsQuery();
   query.prefetch = true;
-  var rows = query.rows();
+  var rows = query.run();
   var data = [];
-  while (row = rows.nextRow()) {
+  while (row = rows.next()) {
     var props = row.documentProperties;
     data.push({
       title: String.format("%s (%s)", props.name, props.symbol)